LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/remove-attr-script-keeps-blocking-expected.txt

Some text

PASS Rendering is blocked before render-blocking resources are loaded
PASS Render-blocking script is loaded and evaluated
